requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
comtypes==1.1.10
decorator==4.4.2
docformatter==1.4
ffmpeg==1.4
flake8==4.0.1
idna==3.3
imageio==2.10.4
imageio-ffmpeg==0.4.5
mccabe==0.6.1
moviepy==1.0.3
numpy==1.21.4
opencv-python==4.5.4.58
Pillow==8.4.0
praw==7.5.0
prawcore==2.3.0
proglog==0.1.9
pycodestyle==2.8.0
pyflakes==2.4.0
pypiwin32==223
pyttsx3==2.90
pywin32==302
requests==2.26.0
toml==0.10.2
tqdm==4.62.3
untokenize==0.1.1
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.1
